+++ release-notes.md
@@ -16,6 +16,7 @@ Changes:
 
 Fixes:
   - Fixed the token-type for medDream in the Quick Viewer button.
+  - Remove the DatePicker icon to provide more room for the text.
 
 
 1.7.1 (2025-01-22)